requirements.txt
Django==1.8.3
MySQL-python==1.2.5
Pillow==2.9.0
easy-thumbnails==2.2
pyquery==1.2.9
unirest==1.1.7
djangorestframework==3.2.2
django-registration-redux==1.2
gunicorn==19.3.0
django-debug-toolbar==1.3.2
django-crispy-forms==1.4.0
drf-extensions==0.2.7
python-memcached==1.57
eventlet==0.17.4
pytz==2015.4
django-htmlmin==0.7.0
celery==3.1.18
newrelic==2.54.0.41
python-dateutil==2.4.2
flower==0.8.3